installrecovery: support MTD partitions

+++ envsetup.sh
@@ -1611,7 +1611,12 @@ function installrecovery()
     if (adb shell cat /system/build.prop | grep -q "ro.cm.device=$CM_BUILD");
     then
         adb push $OUT/recovery.img /cache/
-        adb shell dd if=/cache/recovery.img of=$PARTITION
+        if [ "$PARTITION_TYPE" == "mtd" ];
+        then
+            adb shell flash_image $PARTITION /cache/recovery.img
+        else
+            adb shell dd if=/cache/recovery.img of=$PARTITION
+        fi
         echo "Installation complete."
     else
         echo "The connected device does not appear to be $CM_BUILD, run away!"
